demo_ai_analytics.py
#!/usr/bin/env python
"""
Demo script to showcase the AI Analytics Integration feature
"""

import webbrowser
import time
import os

def demo_ai_analytics():
    """Demo the AI Analytics Integration"""
    
    print("🎬 AI Analytics Integration Demo")
    print("="*50)
    
    print("\n📋 What you'll see:")
    print("1. Board analytics page with comprehensive metrics")
    print("2. New AI Analytics Summary section at the top")
    print("3. 'Generate AI Summary' button")
    print("4. AI-powered insights and recommendations")
    
    print("\n🚀 Features demonstrated:")
    print("✨ Intelligent analysis of task distribution")
    print("📊 Productivity and efficiency insights")
    print("🎯 Lean Six Sigma value stream analysis")
    print("👥 Team workload assessment")
    print("📈 Actionable improvement recommendations")
    
    print("\n🤖 AI provides insights on:")
    print("• Overall project health")
    print("• Bottlenecks and issues")
    print("• Process improvement opportunities")
    print("• Resource allocation recommendations")
    print("• Lean Six Sigma optimization")
    
    # Check if server is running
    print("\n🔍 Checking if development server is running...")
    try:
        import requests
        response = requests.get("http://127.0.0.1:8000/", timeout=3)
        if response.status_code == 200:
            print("✅ Development server is running")
        else:
            print("⚠️ Server responded but may have issues")
    except:
        print("❌ Development server not running")
        print("💡 Start it with: python manage.py runserver")
        return
    
    print("\n🌐 Opening demo in browser...")
    
    # Open the analytics page
    url = "http://127.0.0.1:8000/board/7/analytics/"
    webbrowser.open(url)
    
    print(f"📱 Opened: {url}")
    
    print("\n📖 Demo Instructions:")
    print("1. Log in as admin/adminpass123 if prompted")
    print("2. Look for the blue 'AI Analytics Summary' section at the top")
    print("3. Click 'Generate AI Summary' button")
    print("4. Watch the AI analyze your board data")
    print("5. Review the comprehensive insights provided")
    
    print("\n💡 Key Features to Notice:")
    print("• Real-time AI analysis of your board metrics")
    print("• Structured recommendations by category")
    print("• Lean Six Sigma process efficiency insights")
    print("• Team performance and workload analysis")
    print("• Specific, actionable improvement suggestions")
    
    print("\n🎯 What makes this special:")
    print("• Goes beyond simple metrics to provide insights")
    print("• Tailored recommendations for your specific data")
    print("• Integrates Lean Six Sigma methodology")
    print("• Helps identify bottlenecks and inefficiencies")
    print("• Saves time by summarizing complex analytics")
    
    print("\n" + "="*50)
    print("🎉 Enjoy exploring the AI Analytics feature!")
    print("="*50)

if __name__ == "__main__":
    demo_ai_analytics()
