requirements.txt
Django==5.0.1
django-allauth==0.61.1
django-colorfield==0.11.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.1.0
django-widget-tweaks==1.5.0
python-dotenv==1.0.0
gunicorn==23.0.0
whitenoise==6.6.0
google-generativeai==0.8.5
python-docx==1.1.0
PyPDF2==3.0.1
lxml==5.4.0